Imperative vs Declarative
React is Declarative and Vanilla Js is Imperative.
In imperative program we have to tell how something should be done
meaning we have to describe every step on how to do something in program.
In declarative program we just tell what should be done by program
then program will worry about how to get it done.

For example when we want to have header in our html,
by vanilla js we have to describe step by step how to have it with js
from creating header tag element, setting the text content of the element,
setting any attribute one by one if needed, and appending the element
to DOM. but by React we only need to tell what element we wanna have in our html
then where will it be displayed then react dom will get it done using jsx.