Create other custom components
In react, custom component is like a function that
consists of ui elements in JSX structure that
could be called over and over again
to add it into real DOM.
Its name is not written in camel case but in Capital Cases.
it needs to be nested under single parent element.
It is called by using tag for example:
<Section1 />